Rename rb_experiment.py to standard_rb.py
