+++ qiskit_experiments/library/randomized_benchmarking/__init__.py
@@ -42,7 +42,7 @@ Analysis
 
     RBUtils
 """
-from .rb_experiment import StandardRB
+from .standard_rb import StandardRB
 from .interleaved_rb_experiment import InterleavedRB
 from .rb_analysis import RBAnalysis
 from .interleaved_rb_analysis import InterleavedRBAnalysis
